+++ src/Sequence/Implementation.php
@@ -184,13 +184,13 @@ interface Implementation extends \Countable
     public function pad(int $size, $element): self;
 
     /**
-     * Return a sequence of 2 sequences partitioned according to the given predicate
+     * Return 2 Sequences partitioned according to the given predicate
      *
      * @param callable(T): bool $predicate
      *
-     * @return Map<bool, Sequence<T>>
+     * @return array{Sequence<T>, Sequence<T>}
      */
-    public function partition(callable $predicate): Map;
+    public function partition(callable $predicate): array;
 
     /**
      * Slice the sequence