+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Changed
 
 - `Innmind\Immutable\Map::partition()` now returns an array to allow for destructuring
+- `Innmind\Immutable\Sequence::partition()` now returns an array to allow for destructuring
 
 ### Removed